requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
requests
pydantic
pydantic[email]